refactor: 파일명 변경 (.prefab.ts를 Prefab.ts로)
Timer.prefab.ts라는 .prefab식의 프리펩 파일명은 페이저 에디터에서 꺠지는 문제가 있어 TimerPrefab.ts로 변경함

바꾸면서 모종의 이유로 이미 바꿨는데 아직 소문자로 되어있던 AppleGameManager도 파스칼케이스로 변경
